+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Version 1.0.2
 =============
 
 - Fixed issue #7 with mixed-case options
+- Fixed issue #7 with add_before/add_after problem
 
 Version 1.0.1
 =============